1) Rename route/cruds.go to route/reconcile_resources.go. 2) Move reconcileXXX functions in revision/cruds.go to revision/reconcile_resources.go
